separated classes into different files and folders
