+++ Decorator/Farm-Restaurant/RestaurantDish.cs
@@ -13,52 +13,4 @@ namespace Farm_Restaurant
     {
         public abstract void Display();
     }
-
-
-    /// <summary>
-    /// A ConcreteComponent class
-    /// </summary>
-    class FreshSalad : RestaurantDish
-    {
-        private string _greens;
-        private string _cheese; //I am going to use this pun everywhere I can
-        private string _dressing;
-
-        public FreshSalad(string greens, string cheese, string dressing)
-        {
-            _greens = greens;
-            _cheese = cheese;
-            _dressing = dressing;
-        }
-
-        public override void Display()
-        {
-            Console.WriteLine("\nFresh Salad:");
-            Console.WriteLine(" Greens: {0}", _greens);
-            Console.WriteLine(" Cheese: {0}", _cheese);
-            Console.WriteLine(" Dressing: {0}", _dressing);
-        }
-    }
-
-    /// <summary>
-    /// A ConcreteComponent class
-    /// </summary>
-    class Pasta : RestaurantDish
-    {
-        private string _pastaType;
-        private string _sauce;
-
-        public Pasta(string pastaType, string sauce)
-        {
-            _pastaType = pastaType;
-            _sauce = sauce;
-        }
-
-        public override void Display()
-        {
-            Console.WriteLine("\nClassic Pasta:");
-            Console.WriteLine(" Pasta: {0}", _pastaType);
-            Console.WriteLine(" Sauce: {0}", _sauce);
-        }
-    }
 }
